refactor: 기능 목록 정리 수정에 따른 클래스 및 메서드 이름 변경
ReadMe.md 파일에서 수정된 기능인 자동차 전진 기능에 따라 기존 클래스 및 메서드 네임 변경

변경된 점
- MoveDistanceCalculator -> CarMover 로 이름 변경
