--- src/main/java/racingcar/executor/calculator/distance/MoveDistanceCalculator.java
@@ -1,7 +1,0 @@
-package racingcar.executor.calculator.distance;
-
-import java.util.Map;
-
-public interface MoveDistanceCalculator {
-    Map<String, Integer> calculateDistance(Map<String, Integer> currentHistory);
-}
