+++ src/main/java/racingcar/executor/mover/CarMover.java
@@ -1,0 +1,7 @@
+package racingcar.executor.mover;
+
+import java.util.Map;
+
+public interface CarMover {
+    Map<String, Integer> run(Map<String, Integer> currentHistory);
+}
